+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Journal des modifications
 5.2.5
 -----
 - [FIX] Erreur dans la recherche de doctroants par établissement et ED (plantage impactant les listes de diff).
+- [FIX] Recherche de rapports d'activités : restrictions par ED/UR selon le rôle endossé et ses privilèges.
 
 5.2.4
 -----